pyproject.toml
[tool.poetry]
name = "algo-neoquantxperience"
version = "0.1.0"
description = ""
authors = ["ma_durov <mdurov@neoflex.ru>"]
readme = "README.md"
packages = [{include = "algo_neoquantxperience"}]

[tool.poetry.dependencies]
python = "^3.9"
moexalgo = "^1.0.1"
pandas = "^2.1.3"
notebook = "^7.0.6"
langchain = "^0.0.345"
gigachat = "^0.1.8"
telethon = "^1.33.0"
python-dotenv = "^1.0.0"
emoji = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
